https://jira.aplana.com/browse/SBRFACCTAX-16175 Исправил косяк по прошлому коммиту

+++ dao/src/main/java/com/aplana/sbrf/taxaccounting/dao/impl/datarow/DataRowDaoImpl.java
@@ -89,7 +89,7 @@ public class DataRowDaoImpl extends AbstractDao implements DataRowDao {
 
 		boolean isNumber = false;
 		try {
-			Integer.parseInt(key);
+			Double.parseDouble(key);
 			isNumber = true;
 		} catch (NumberFormatException ignored) {}
